% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{basicranial_bones}
\alias{basicranial_bones}
\title{Fetal Basicranial Bones Data}
\format{
A data frame containing measurements of fetal basicranial bones
}
\source{
Fazekas, I.G. and Kósa, F. (1978) Forensic Fetal Osteology. Akadémiai Kiadó, Budapest.
}
\usage{
basicranial_bones
}
\description{
Dataset containing measurements of fetal basicranial bones from Fazekas and Kosa
}
\keyword{datasets}
