% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{basocranial_bones}
\alias{basocranial_bones}
\title{Basicranial Bone Measurements}
\format{
A data frame with 138 rows and 17 variables:
\describe{
\item{Lunar_month}{Gestational age in lunar months}
\item{Sex}{Sex of the fetus}
\item{No_of_cases}{Case number}
\item{Order}{Order of cases}
\item{Body_length}{Body length in cm}
\item{Body_weight}{Body weight in g}
\item{Greater_wing_bone_sphenoid}{Greater wing bone of sphenoid measurements}
...
}
}
\source{
Fazekas and Kósa dataset
}
\usage{
basocranial_bones
}
\description{
A dataset containing measurements of basocranial bones in fetal specimens
}
\keyword{datasets}
