% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{extremity_dimensions}
\alias{extremity_dimensions}
\title{Extremity Dimensions}
\format{
A data frame with 138 rows and 14 variables:
\describe{
\item{Lunar_month}{Gestational age in lunar months}
\item{Sex}{Sex of the fetus}
\item{No_of_cases}{Case number}
\item{Order}{Order of cases}
\item{Body_length}{Body length in cm}
\item{Body_weight}{Body weight in g}
\item{Humerus_Length}{Length of humerus in mm}
...
}
}
\source{
Fazekas and Kósa dataset
}
\usage{
extremity_dimensions
}
\description{
A dataset containing measurements of extremity dimensions in fetal specimens
}
\keyword{datasets}
