% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{facial_bones}
\alias{facial_bones}
\title{Facial Bone Measurements}
\format{
A data frame with 138 rows and 19 variables:
\describe{
\item{Lunar_month}{Gestational age in lunar months}
\item{Sex}{Sex of the fetus}
\item{No_of_cases}{Case number}
\item{Order}{Order of cases}
\item{Body_length}{Body length in mm}
\item{Body_weight}{Body weight in g}
\item{Nasal_bone_length}{Length of nasal bone in mm}
...
}
}
\source{
Fazekas and Kósa dataset
}
\usage{
facial_bones
}
\description{
A dataset containing measurements of facial bones in fetal specimens
}
\keyword{datasets}
