% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{liverpool_bones}
\alias{liverpool_bones}
\title{Liverpool Bone Length Collection}
\format{
A data frame with rows and 17 variables:
\describe{
\item{Specimen_Number}{Specimen identifier}
\item{Age_gestational}{Gestational age}
\item{Sex}{Sex of the fetus}
\item{L_Humerus}{Left humerus length in mm}
...
}
}
\source{
Liverpool Collection
}
\usage{
liverpool_bones
}
\description{
A dataset containing bone length measurements from Liverpool collection
}
\keyword{datasets}
