% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{liverpool_lengths}
\alias{liverpool_lengths}
\title{Liverpool Collection Long Bone Lengths}
\format{
A data frame containing fetal long bone measurements
}
\source{
Liverpool Collection data collected by Thomas O'Mahoney
}
\usage{
liverpool_lengths
}
\description{
Dataset containing long bone measurements from the Liverpool Collection
}
\keyword{datasets}
