% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ribs}
\alias{ribs}
\title{Fetal Ribs Data}
\format{
A data frame containing measurements of fetal ribs
}
\source{
Fazekas, I.G. and Kósa, F. (1978) Forensic Fetal Osteology. Akadémiai Kiadó, Budapest.
}
\usage{
ribs
}
\description{
Dataset containing measurements of fetal ribs from Fazekas and Kosa
}
\keyword{datasets}
