% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shoulder_pelvic}
\alias{shoulder_pelvic}
\title{Shoulder and Pelvic Girdle Measurements}
\format{
A data frame with rows and variables:
\describe{
\item{Lunar_month}{Gestational age in lunar months}
\item{Sex}{Sex of the fetus}
\item{No_of_cases}{Case number}
\item{Order}{Order of cases}
\item{Body_length}{Body length in cm}
\item{Body_weight}{Body weight in g}
\item{Clavicle_length}{Length of clavicle in mm}
...
}
}
\source{
Fazekas and Kósa dataset
}
\usage{
shoulder_pelvic
}
\description{
A dataset containing measurements of shoulder and pelvic girdle bones
}
\keyword{datasets}
