% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{small_bones}
\alias{small_bones}
\title{Fetal Small Bones Data}
\format{
A data frame containing measurements of fetal small bones
}
\source{
Fazekas, I.G. and Kósa, F. (1978) Forensic Fetal Osteology. Akadémiai Kiadó, Budapest.
}
\usage{
small_bones
}
\description{
Dataset containing measurements of fetal small bones from Fazekas and Kosa
}
\keyword{datasets}
