% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{squamous_occipital}
\alias{squamous_occipital}
\title{Fetal Squamous Occipital Data}
\format{
A data frame containing measurements of fetal squamous occipital
}
\source{
Grzonkowska et al. publication details
}
\usage{
squamous_occipital
}
\description{
Dataset containing measurements of fetal squamous occipital from "Grzonkowska, M., Baumgart, M., Badura, M., Wiśniewski, M. and Szpinda, M., 2021. Quantitative anatomy of the fused ossification center of the occipital squama in the human fetus. Plos one, 16(2), p.e0247601. https://doi.org/10.1371/journal.pone.0247601"
}
\keyword{datasets}
