% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{squamous_temporal}
\alias{squamous_temporal}
\title{Fetal Squamous Temporal Data}
\format{
A data frame containing measurements of fetal squamous temporal
}
\source{
Grzonkowska et al. publication details
}
\usage{
squamous_temporal
}
\description{
Dataset containing measurements of fetal squamous temporal from Grzonkowska et al.
}
\keyword{datasets}
