% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vault_bones}
\alias{vault_bones}
\title{Fetal Vault Bones Data}
\format{
A data frame containing measurements of fetal vault bones
}
\source{
Fazekas, I.G. and Kósa, F. (1978) Forensic Fetal Osteology. Akadémiai Kiadó, Budapest.
}
\usage{
vault_bones
}
\description{
Dataset containing measurements of fetal vault bones from Fazekas and Kosa
}
\keyword{datasets}
