% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{extremity_dimensions}
\alias{extremity_dimensions}
\title{Fetal Extremity Dimensions Data}
\format{
A data frame containing measurements of fetal extremity dimensions
}
\source{
Fazekas, I.G. and Kósa, F. (1978) Forensic Fetal Osteology. Akadémiai Kiadó, Budapest.
}
\usage{
extremity_dimensions
}
\description{
Dataset containing measurements of fetal extremity dimensions from Fazekas and Kosa
}
\keyword{datasets}
