% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ribs}
\alias{ribs}
\title{Rib Measurements}
\format{
A data frame with 138 rows and variables:
\describe{
\item{Lunar_month}{Gestational age in lunar months}
\item{Sex}{Sex of the fetus}
\item{No_of_cases}{Case number}
\item{Order}{Order of cases}
\item{Body_length}{Body length in cm}
\item{Body_weight}{Body weight in g}
\item{Rib_I}{First rib measurement in mm}
...
}
}
\source{
Fazekas and Kósa dataset
}
\usage{
ribs
}
\description{
A dataset containing measurements of ribs in fetal specimens
}
\keyword{datasets}
