% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shoulder_pelvic}
\alias{shoulder_pelvic}
\title{Fetal Shoulder and Pelvic Girdle Data}
\format{
A data frame containing measurements of fetal shoulder and pelvic girdle
}
\source{
Fazekas, I.G. and Kósa, F. (1978) Forensic Fetal Osteology. Akadémiai Kiadó, Budapest.
}
\usage{
shoulder_pelvic
}
\description{
Dataset containing measurements of fetal shoulder and pelvic girdle from Fazekas and Kosa
}
\keyword{datasets}
