% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{small_bones}
\alias{small_bones}
\title{Small Bone Measurements}
\format{
A data frame with rows and variables:
\describe{
\item{Lunar_month}{Gestational age in lunar months}
\item{Sex}{Sex of the fetus}
\item{No_of_cases}{Case number}
\item{Order}{Order of cases}
\item{Body_length}{Body length in mm}
\item{Body_weight}{Body weight in mm}
\item{Atlas_length}{Length of atlas in mm}
...
}
}
\source{
Fazekas and Kósa dataset
}
\usage{
small_bones
}
\description{
A dataset containing measurements of small bones in fetal specimens
}
\keyword{datasets}
