% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{squamous_occipital}
\alias{squamous_occipital}
\title{Fetal Squamous Occipital Data}
\format{
A data frame containing measurements of fetal squamous occipital
}
\source{
Grzonkowska et al. publication details
}
\usage{
squamous_occipital
}
\description{
Dataset containing measurements of fetal squamous occipital from Grzonkowska et al.
}
\keyword{datasets}
