% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{squamous_temporal}
\alias{squamous_temporal}
\title{Fetal Squamous Temporal Data}
\format{
A data frame containing measurements of fetal squamous temporal
}
\source{
Grzonkowska et al. publication details
}
\usage{
squamous_temporal
}
\description{
Dataset containing measurements of fetal squamous temporal from Grzonkowska, M., Baumgart, M., Kułakowski, M. and Szpinda, M., 2023. Quantitative anatomy of the primary ossification center of the squamous part of temporal bone in the human fetus. Plos one, 18(12), p.e0295590. https://doi.org/10.1371/journal.pone.0295590 
}
\keyword{datasets}
